Fix errors missed while pulling images
On the /images/pull endpoint, podman's StatusOK response does
not warranty that the image was actually pulled, just that
the request was well formatted and no InteralServerError happened.
If any error happens during the download, the error is written to
the output stream in the error field.

This error could be identified in the test from the previous commit