driver: fix image_pull_timeout
Previously the value for `image_pull_timeout` was being capped by
`client_http_timeout` because the image pull request used the same HTTP
client as other requests.

Increasing `client_http_timeout` in order to be able to set a higher
`image_pull_timeout` is not a good option since it requires updating all
client's configuration. It also affects all other HTTP requests, which
may not be desirable.

The change creates a separate client for slow operations that don't have
a timeout configured. Operation timeout is controlled using a
ContextWithTimeout().