config: set recover_stopped to default to false (#260)
The use of recover_stopped may cause the Nomad agent to hang on startup,
as the plugin tries to start an exited podman task. Podman itself will
hang forever in this state, and the http client on the Nomad side is also
unable to timeout in this case. The result is a permenantly hung Nomad
agent, until someone force kills either Nomad or Podman.

Also emit a log warning that recover_stopped should not be used. We leave
it in place for compatability.

Fixes #229