Resolve #293 and #358 adding ipv6_address, ipv4_address, static_ips and static_mac options for podman tasks.
Podman API version checks whether to set the properties directly or whether to use the new PerNetworkOptions framework.

ipv6_address and ipv4_address are provided for backwards compatibility, and for compatibility with docker. For Podman 4.0.0 these are merged into static_ips and sent as a PerNetorkOptions entry for the default network.
Added API version check, static mac support, new array-based option for static IPs to match the podman API