Support for multiple podman sockets (#371)
The current driver only allows you to specify one rootless socket in the
driver configuration. Having different host users run containers makes sure that
the container UIDs don't map to the same user on the host.

Add support for multiple "socket" blocks in the driver configuration. The Tasks
can reference a socket using the name of the socket. As far as I have manually
tested I have made it backwards compatible (but mutually exclusive) to be
combined with the existing `socket_path`. I've added some code so there's always
a default podman client option even if it's not specified by the user
(preferably this is tackled in the HCL parsing but I didn't find how I could it
handle this as an error early in the spec parsing).

Fixes: #284
Fixes: #84