user namespace: add required ID mappings field when specifying auto
When using a private user namespace, we need to include the `IDMappings` block
in the create container's request. This block includes UID and GID mappings
parsed out of the string the user gives us. Without it, specifying an `"auto"`
configuration silently fails to wire up the correct mappings.

This implementation doesn't include additional UID/GID mappings, as we don't
have that configuration option.

Fixes: https://github.com/hashicorp/nomad-driver-podman/issues/470